Substitue AMI creation by s3+ebs (partial)

+++ scripts/remote_setup.sh
@@ -3,17 +3,23 @@ set -e
 
 cd /home/ubuntu
 
-wget -q https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
-bash Miniconda3-latest-Linux-x86_64.sh -b -p /home/ubuntu/miniconda3
-rm Miniconda3-latest-Linux-x86_64.sh
-echo "source /home/ubuntu/miniconda3/etc/profile.d/conda.sh" >> .bashrc
-source /home/ubuntu/miniconda3/etc/profile.d/conda.sh
-conda create -y -q -n nimbo python=3.7 
-conda activate nimbo
-conda env update -q --file nimbo-environment.yml
-conda install -q -c conda-forge awscli -y
+CONDASH=/home/ubuntu/miniconda3/etc/profile.d/conda.sh
+if [ -f "$CONDASH" ]; then
+    conda activate nimbo
+    conda env update -q --file nimbo-environment.yml
+else
+    wget -q https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
+    bash Miniconda3-latest-Linux-x86_64.sh -b -p /home/ubuntu/miniconda3
+    rm Miniconda3-latest-Linux-x86_64.sh
+    echo "source /home/ubuntu/miniconda3/etc/profile.d/conda.sh" >> .bashrc
+    source /home/ubuntu/miniconda3/etc/profile.d/conda.sh
+    conda env create -q -n nimbo --file nimbo-environment.yml
+    conda activate nimbo
+    conda install -q -c conda-forge awscli -y
+fi
 
-mkdir -p /home/ubuntu/data/datasets
-aws s3 cp --recursive s3://nimbo-main-bucket/data/datasets /home/ubuntu/data/datasets
+# Import everything from bucket. At some point the env, code, datasets and results folders should be imported separately
+mkdir -p data/datasets
+mkdir -p data/results
 
-rm -rf /home/ubuntu/data/datasets
+aws s3 cp s3://nimbo-main-bucket /home/ubuntu
